more refactors: - Move common raft attribute reader and writer methods to a common file - Rafactor appendentries rpc to acquire lock at the start of the request   processing - Dont update follower commit index to leaders commit index, instead use   min(leader_commit_index, leaderPrevLogIndex + len(incoming log entries)) - While transitioning a raft server from within an RPC, release and   reacquire the lock so that background go-routines are not blocked.
