+++ raft_rpcs.go
@@ -50,7 +50,7 @@ func (rf *Raft) RequestVote(args *RequestVoteArgs, reply *RequestVoteReply) {
 		rf.persist()
 		if rf.IsLeader() {
 			rf.Log("stepping down as leader from request vote")
-			rf.TransitionToFollower(fmt.Sprintf("from request vote [%s]from leader with term: %d", rf.candidateID, rf.currentTerm))
+			rf.SafeTransitionWithMutex(rf.TransitionToFollower, fmt.Sprintf("from request vote [%s]from leader with term: %d", rf.candidateID, rf.currentTerm))
 		} else if rf.IsFollower() {
 			rf.blockingNotifyWhile(rf.heartBeatChan, "sending heartbeat from request vote", rf.IsFollower)
 		}
@@ -68,7 +68,7 @@ func (rf *Raft) RequestVote(args *RequestVoteArgs, reply *RequestVoteReply) {
 		rf.persist()
 		if rf.IsLeader() {
 			rf.Log("stepping down as leader from request vote")
-			rf.TransitionToFollower(fmt.Sprintf("from request vote [%s]from leader with term: %d", rf.candidateID, rf.currentTerm))
+			rf.SafeTransitionWithMutex(rf.TransitionToFollower, fmt.Sprintf("from request vote [%s]from leader with term: %d", rf.candidateID, rf.currentTerm))
 		}
 	}
 	rf.Log("rejected request vote from %s for term %d. updated term: %d", args.CandidateID, args.Term, rf.currentTerm)
@@ -96,10 +96,12 @@ func (rf *Raft) candidateUptodate(candidateLastLogIndex int32, candidateLastLogT
 }
 
 func (rf *Raft) AppendEntries(args *AppendEntriesArgs, reply *AppendEntriesReply) {
-	if args.Term >= rf.GetCurrentTerm() {
+	rf.mu.Lock()
+	defer rf.mu.Unlock()
+	if args.Term >= rf.currentTerm {
 		rf.acknowledgeHeartBeat(args)
 		if len(args.Entries) == 0 {
-			rf.UpdateCommitIndex(args.LeaderCommit)
+			rf.updateCommitIndex(args.LeaderCommit)
 			// If its a heartbeat only append entry rpc call.
 			*reply = AppendEntriesReply{
 				Success: true,
@@ -109,14 +111,14 @@ func (rf *Raft) AppendEntries(args *AppendEntriesArgs, reply *AppendEntriesReply
 		}
 		success, logChanged, extraInfo := rf.appendToLog(args)
 		if logChanged {
-			rf.Persist()
+			rf.persist()
 		}
 		if success {
 			*reply = AppendEntriesReply{
 				Success: true,
 				Term:    args.Term,
 			}
-			rf.UpdateCommitIndex(args.LeaderCommit)
+			rf.updateCommitIndex(min(args.LeaderCommit, args.PrevLogIndex+len(args.Entries)))
 		} else {
 			*reply = AppendEntriesReply{
 				Success:   false,
@@ -128,13 +130,11 @@ func (rf *Raft) AppendEntries(args *AppendEntriesArgs, reply *AppendEntriesReply
 	}
 	*reply = AppendEntriesReply{
 		Success: false,
-		Term:    rf.GetCurrentTerm(),
+		Term:    rf.currentTerm,
 	}
 }
 
 func (rf *Raft) appendToLog(args *AppendEntriesArgs) (bool, bool, *AppendEntryExtras) {
-	rf.mu.Lock()
-	defer rf.mu.Unlock()
 	lastEntryFromLeader := args.Entries[len(args.Entries)-1]
 	currentLastLogIndex := len(rf.logContents)
 	currentLogLen := currentLastLogIndex
@@ -152,8 +152,9 @@ func (rf *Raft) appendToLog(args *AppendEntriesArgs) (bool, bool, *AppendEntryEx
 			rf.logContents = append(rf.logContents, args.Entries...)
 			return true, true, nil
 		}
-		// If the entry at this peers log at logEntryFromLeader.LogIndex doesnt match with the
-		// logEntryFromLeader.
+
+		// Only truncate the logs if the incoming entries modified the servers log.
+		// Otherwise and out of order request can wipe out legit entries from the servers log.
 		if logModified := rf.InsertEntries(currentLogLen, args); logModified {
 			rf.logContents = rf.logContents[:lastEntryFromLeader.LogIndex]
 			return true, true, nil
@@ -173,6 +174,8 @@ func (rf *Raft) appendToLog(args *AppendEntriesArgs) (bool, bool, *AppendEntryEx
 		// If the prev entry in the peers log and leaders prev entry match.
 		if currentPrevLogEntry.LogIndex == args.PrevLogIndex && currentPrevLogEntry.Term == args.PrevLogTerm {
 			modified := rf.InsertEntries(currentLogLen, args)
+			// Only truncate the logs if the incoming entries modified the servers log.
+			// Otherwise and out of order request can wipe out legit entries from the servers log.
 			if modified {
 				rf.logContents = rf.logContents[:lastEntryFromLeader.LogIndex]
 			}
@@ -218,16 +221,16 @@ func (rf *Raft) acknowledgeHeartBeat(args *AppendEntriesArgs) {
 	if rf.IsFollower() {
 		rf.blockingNotifyWhile(rf.heartBeatChan, "acking heart beat from leader", rf.IsFollower)
 	}
-	if args.Term > rf.GetCurrentTerm() {
-		rf.SetTerm(args.Term)
+	if args.Term > rf.currentTerm {
+		rf.setTerm(args.Term)
 	}
 
 	if rf.IsElectionRunning() {
 		rf.Log("asked to stop election and tranistioning to follower")
-		rf.TransitionToFollower(fmt.Sprintf("from ack heartbeat [%s]from candidate with term: %d", rf.candidateID, rf.currentTerm))
+		rf.SafeTransitionWithMutex(rf.TransitionToFollower, fmt.Sprintf("from ack heartbeat [%s]from candidate with term: %d", rf.candidateID, rf.currentTerm))
 	}
 	if rf.IsLeader() {
-		rf.TransitionToFollower(fmt.Sprintf("from ack heartbeat [%s]from leader with term: %d", rf.candidateID, rf.currentTerm))
+		rf.SafeTransitionWithMutex(rf.TransitionToFollower, fmt.Sprintf("from ack heartbeat [%s]from leader with term: %d", rf.candidateID, rf.currentTerm))
 		rf.Log("asked to transition from leader to follower")
 	}
 
@@ -238,11 +241,17 @@ func (rf *Raft) acknowledgeHeartBeat(args *AppendEntriesArgs) {
 }
 
 func (rf *Raft) UpdateCommitIndex(leaderCommitIndex int) {
-	currentLogLen := rf.LogLength()
 	rf.mu.Lock()
+	defer rf.mu.Unlock()
+	currentLogLen := rf.logLength()
 	rf.commitIndex = min(currentLogLen, leaderCommitIndex)
 	rf.ApplyToStateMachine()
-	rf.mu.Unlock()
+}
+
+func (rf *Raft) updateCommitIndex(leaderCommitIndex int) {
+	currentLogLen := rf.logLength()
+	rf.commitIndex = min(currentLogLen, leaderCommitIndex)
+	rf.ApplyToStateMachine()
 }
 
 func min(x, y int) int {
